pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fraud-platform"
version = "0.1.0"
description = "Production-minded real-time fraud detection platform monorepo."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "OpenAI Codex" }
]
dependencies = [
  "numpy>=1.26.4,<3.0.0",
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "pydantic>=2.8.0,<3.0.0",
  "pydantic-settings>=2.4.0,<3.0.0",
  "sqlalchemy>=2.0.32,<3.0.0",
  "alembic>=1.13.2,<2.0.0",
  "psycopg[binary]>=3.2.1,<4.0.0",
  "kafka-python-ng>=2.2.3,<3.0.0",
  "redis>=5.0.8,<6.0.0",
  "httpx>=0.27.0,<1.0.0",
  "prometheus-client>=0.20.0,<1.0.0",
  "python-json-logger>=2.0.7,<4.0.0",
  "orjson>=3.10.7,<4.0.0",
  "pandas>=2.2.2,<3.0.0",
  "pyyaml>=6.0.2,<7.0.0",
]

[project.optional-dependencies]
api = [
  "python-multipart>=0.0.9,<1.0.0",
]
producer = [
  "faker>=28.0.0,<40.0.0",
]
stream-worker = [
  "bytewax>=0.19.1,<1.0.0",
  "confluent-kafka>=2.5.0,<3.0.0",
]
runtime = [
  "mlflow>=2.16.0,<3.0.0",
  "joblib>=1.4.2,<2.0.0",
  "xgboost>=2.1.1,<3.0.0",
]
trainer = [
  "evidently>=0.6.1,<1.0.0",
  "polars>=1.6.0,<2.0.0",
  "scikit-learn>=1.5.1,<2.0.0",
  "shap>=0.46.0,<1.0.0",
]
dev = [
  "mypy>=1.11.1,<2.0.0",
  "pre-commit>=3.8.0,<4.0.0",
  "pytest>=8.3.2,<9.0.0",
  "pytest-asyncio>=0.24.0,<1.0.0",
  "pytest-cov>=5.0.0,<6.0.0",
  "respx>=0.21.1,<1.0.0",
  "ruff>=0.6.3,<1.0.0",
]

[project.scripts]
fraud-api = "fraud_platform_api.main:run"
fraud-producer = "fraud_platform_producer.main:run"
fraud-stream-worker = "fraud_platform_stream_worker.main:run"
fraud-trainer = "fraud_platform_trainer.main:run"
fraud-producer-cli = "fraud_platform_producer.cli:main"
fraud-stream-runner = "fraud_platform_stream_worker.runner:main"
fraud-trainer-cli = "fraud_platform_trainer.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = [
  "apps/api/src",
  "apps/producer/src",
  "apps/stream-worker/src",
  "apps/trainer/src",
  "libs/common/src",
  "libs/contracts/src",
  "libs/feature_engineering/src",
  "libs/feature_store/src",
  "libs/model_runtime/src",
  "libs/observability/src",
  "libs/persistence/src",
  "libs/rules/src",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = [
  "apps",
  "libs",
  "tests",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = []
